Added ID system for set-members.
Changed the system to assign each set-member an integer ID, which is now used instead of the member's value for the process of finding all the possible permutations. This new system avoids conflicts between repeating values in the set and the check to avoid placing the same member in a permutation more than once.