Cleanup and upgrades
The first iteration of this project was really rough. I was and still am learning...
This new version changes things so that there is an importable header for the permutation solving functions. The new header includes overloads for different input data types.
There is now also an included EXE to act as a command-line tool or be run as a subprocess.